feat: Delete images
