+++ src/modules/Article/views/CreateOrEdit/Article/components/GalleryImages/Index.vue
@@ -1,20 +1,29 @@
 <script setup lang="ts">
-import { ref, watch, computed } from 'vue';
+import { ref, watch, inject } from 'vue';
 import ModalShowPreview from './ModalShowPreview.vue';
-
-const startOfImagePath: string = import.meta.env.VITE_PRODUCT_API_URL
+import type { ImageType, Base64 } from "@/modules/Article/types/Image";
+import AppButton from '@/core/components/AppButton.vue';
 
 const props = defineProps<{
   imagePaths: string[]
+  base64Images: Base64<ImageType>[]
 }>()
+
+const {
+  removeImagePath,
+  removeAllImagePaths
+}: {
+  removeImagePath: (index: number) => void,
+  removeAllImagePaths: () => void
+} = inject('image-paths');
 
 const selectImage = ref('')
 
 const previewImage = ref('')
 
-const showPreview = (image: string) => {
+const showPreview = (image: Base64<ImageType>) => {
   selectImage.value = image;
-  previewImage.value = `${startOfImagePath}/${image}`;
+  previewImage.value = image;
 }
 
 const closePreview = () => {
@@ -23,13 +32,13 @@ const closePreview = () => {
 }
 
 watch(
-  ()=> props.imagePaths[0],
+  ()=> props.base64Images[0],
   newValue => { showPreview(newValue) },
   // { once: true }
 )
 
 const classImages = (index: number) =>
-  props.imagePaths[index] === selectImage.value
+  props.base64Images[index] === selectImage.value
     ? 'image-selected'
     : 'image-unselected'
 </script>
@@ -42,20 +51,26 @@ const classImages = (index: number) =>
       @closePreview="closePreview"
     />
     <div class="image-container">
-      <!--a class="prev" onclick="plusSlides(-1)">&#10094;</a-->
       <div
         :class="classImages(index)"
-        v-for="(image, index)  in props.imagePaths"
+        v-for="(image, index)  in base64Images"
         :key="index"
       >
         <img          
-          :src="`${startOfImagePath}/${image}`"
+          :src="image"
           alt="Image" 
           @click="showPreview(image)"
+          @dblclick="removeImagePath(index)"
         />
       </div>
-      <!--a class="next" onclick="plusSlides(1)">&#10095;</a-->
     </div>
+    <AppButton
+      v-if="base64Images.length"
+      type="button"
+      text="Eliminar todas"
+      class="btn btn-danger"
+      @click="removeAllImagePaths()"
+    />
   </div>
 </template>
